# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Always_date_api documentation",
        "description": "Always date documentation for back-end",
        "contact": {
            "name": "FlokY#0101"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/match": {
            "get": {
                "tags": [
                    "Match log"
                ],
                "summary": "Atgriež sakritības ierakstus",
                "description": "Atgriež visus sakritības ierakstus",
                "operationId": "getMatchLog",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Match log"
                ],
                "summary": "Izveido vai atjauno jaunu sakritības ierakstu",
                "description": "Izveido vai atjauno jaunu sakritības ierakstu",
                "operationId": "postMatchLog",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "user_2"
                                ],
                                "properties": {
                                    "user_2": {
                                        "description": "Lietotājs kuram tiks likts vērtējums",
                                        "format": "integer"
                                    },
                                    "mark": {
                                        "description": "Vērtējums 1 - true, 0 - false",
                                        "format": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Vienu reizi var likt vērtējumu."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/match/{id}": {
            "get": {
                "tags": [
                    "Match log"
                ],
                "summary": "Iegūst sakritības ierakstu",
                "description": "Iegūst konkrētu sakritības ierakstu pēc id",
                "operationId": "getMatchLogById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "MatchLog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Match log"
                ],
                "summary": "Izdzēš sakritības ierakstu",
                "description": "Izdzēš sakritības ierakstu pēc id",
                "operationId": "deleteMatchLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "MatchLog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MatchLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/report_log": {
            "get": {
                "tags": [
                    "Report log"
                ],
                "summary": "Iegūst nosūtītās sūdzības",
                "description": "Iegūst visas nosūtītās sūdzibas",
                "operationId": "getReportLog",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Report log"
                ],
                "summary": "Izveido sūdzības par lietotāju",
                "description": "Izveido jaunu sūdzību par kādu lietotāju",
                "operationId": "postReportLog",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "report_type"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "description": "Lietotāja id par kuru tiks nosūtīta sūdzība",
                                        "format": "integer"
                                    },
                                    "report_type": {
                                        "description": "Sūdzības veida id",
                                        "format": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    },
                    "401": {
                        "description": "Nevar nosūtīt sūdzību par sevi."
                    },
                    "402": {
                        "description": "Nevar nosūtīt vairāk par vienu sūdzību."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/report_log/{id}": {
            "get": {
                "tags": [
                    "Report log"
                ],
                "summary": "Iegūst nosūtīto sūdzību",
                "description": "Iegūst konkrēto nosūtīto sūdzību par lietotāju",
                "operationId": "getReportLogById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReportLog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Report log"
                ],
                "summary": "Izdzēš sūdzību par lietotāju",
                "description": "Izdzēs sūdzību par lietotāju pēc id",
                "operationId": "deleteReportLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReportLog id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportLogResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MatchLogResource": {
                "title": "Match log resource",
                "description": "Match log resource",
                "xml": {
                    "name": "MatchLogResource"
                }
            },
...
            "ReportLogResource": {
                "title": "Report log resource",
                "description": "Report log resource",
                "xml": {
                    "name": "ReportLogResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}